API/requirements.txt
torch
fastapi
torchvision
pillow